extract code for adding image to separate component
